requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.4
Pygments==2.4.2
python-dotenv==0.10.3
pytz==2018.3
whitenoise==4.1.4